pyproject.toml
[tool.poetry]
name = "drf-kyrsovaya-7"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
black = "^24.8.0"
isort = "^5.13.2"
ipython = "^8.26.0"
python-dotenv = "^1.0.1"
django = "^5.0.7"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"
redis = "^5.0.8"
psycopg2-binary = "^2.9.9"
django-celery-beat = "^2.6.0"
celery = "^5.4.0"
django-filter = "^24.3"
drf-yasg = "^1.21.7"
stripe = "^10.6.0"
eventlet = "^0.36.1"
djangorestframework-simplejwt = "^5.3.1"
setuptools = "^72.1.0"
django-cors-headers = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
